Race Control Card Fixes
